+++ internal/client/transport.go
@@ -18,8 +18,8 @@ package client
 import (
 	"net/http"
 
+	"github.com/mia-platform/miactl/internal/logger"
 	"github.com/mia-platform/miactl/internal/transport"
-	"github.com/mia-platform/miactl/internal/util"
 	"golang.org/x/oauth2"
 )
 
@@ -60,7 +60,7 @@ func transportForConfig(config *Config) (http.RoundTripper, error) {
 			Insecure: config.Insecure,
 			CAFile:   config.CAFile,
 		},
-		Verbose: util.LogLevel >= 5,
+		Verbose: logger.LogLevel >= 5,
 	}
 
 	if authProvider != nil {