+++ internal/cmd/root.go
@@ -22,7 +22,7 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/mia-platform/miactl/internal/clioptions"
 	"github.com/mia-platform/miactl/internal/cmd/deploy"
-	"github.com/mia-platform/miactl/internal/util"
+	"github.com/mia-platform/miactl/internal/logger"
 	"github.com/spf13/cobra"
 )
 
@@ -38,7 +38,7 @@ func NewRootCommand() *cobra.Command {
 	// initialize clioptions and setup during initialization
 	options := clioptions.NewCLIOptions()
 
-	logger := util.NewLogger(os.Stderr)
+	logger := logger.NewLogger(os.Stderr)
 	rootCmd.SetContext(logr.NewContext(context.Background(), logger))
 
 	// add cmd flags